+++ internal/storage/storage.go
@@ -13,11 +13,11 @@ type Storage interface {
 	GetUserID(ctx context.Context, login string) (uuid.UUID, error)
 	GetOrder(ctx context.Context, orderID models.OrderID) (*models.Order, error)
 	AddOrder(ctx context.Context, userID uuid.UUID, status models.OrderStatus, orderID models.OrderID) error
-	UpdateOrder(ctx context.Context, orderID models.OrderID, status models.OrderStatus, accrual int) error
+	UpdateOrder(ctx context.Context, orderID models.OrderID, status models.OrderStatus, accrual float32) error
 	GetUserOrders(ctx context.Context, userID uuid.UUID) ([]models.Order, error)
 	GetUnfinishedOrderIDs(ctx context.Context) ([]models.OrderID, error)
-	GetUserAccrualBalance(ctx context.Context, userID uuid.UUID) (int, error)
-	AddWithdrawal(ctx context.Context, userID uuid.UUID, orderID models.OrderID, amount int) error
-	GetUserWithdrawalSum(ctx context.Context, userID uuid.UUID) (int, error)
+	GetUserAccrualBalance(ctx context.Context, userID uuid.UUID) (float32, error)
+	AddWithdrawal(ctx context.Context, userID uuid.UUID, orderID models.OrderID, amount float32) error
+	GetUserWithdrawalSum(ctx context.Context, userID uuid.UUID) (float32, error)
 	GetUserWithdrawals(ctx context.Context, userID uuid.UUID) ([]models.Withdrawal, error)
 }
